% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pals-package.R
\docType{package}
\name{palsr-package}
\alias{palsr}
\alias{palsr-package}
\title{palsr: Projected Actor Locations for Spatial Interaction Modeling}
\description{
Implements the Projected Actor Locations (PALS) method for spatial modeling of dyadic interactions between geographically mobile actors, as described in Kim, Liu and Desmarais (2023) \doi{10.1017/psrm.2022.6}. PALS applies exponential-smoothing weights to the spatiotemporal histories of a focal actor and its interaction partners ("alters") to project the location of future interactions. The package provides projection, maximum-similarity parameter estimation by minimizing great-circle (Haversine) prediction error, nonparametric bootstrap with multiple-imputation (Rubin's Rules) pooling, dyadic distance covariate construction, visualization, and a simulated example dataset of subnational conflict.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bdesmarais/palsr}
  \item Report bugs at \url{https://github.com/bdesmarais/palsr/issues}
}

}
\author{
\strong{Maintainer}: Bruce A. Desmarais \email{bruce.desmarais@gmail.com}

Authors:
\itemize{
  \item Bruce A. Desmarais \email{bruce.desmarais@gmail.com}
  \item Sangyeon Kim
  \item Howard Liu
}

}
\keyword{internal}
